# frontend/public/openapi.yaml
openapi: 3.0.3
info:
  title: Trade Nexus API
  version: 1.0.0
  description: |
    Comprehensive API for Trade Nexus - CFTC Position Limits Compliance Platform

    ## Overview
    Trade Nexus provides a complete suite of APIs for managing commodity trading position limits,
    risk management, compliance reporting, and regulatory submissions to the CFTC.

    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain a token via `/api/auth/login`.

    ## Rate Limiting
    - Standard users: 100 requests per minute
    - API keys: 1000 requests per minute

    ## Base URLs
    - **Production**: https://trade-nexus-api.metabilityllc1.workers.dev
    - **Development**: https://trade-nexus-api-dev.metabilityllc1.workers.dev

  contact:
    name: Trade Nexus Support
    email: support@tradenexus.com
  license:
    name: Proprietary
servers:
  - url: https://trade-nexus-api-dev.metabilityllc1.workers.dev
    description: Development server
  - url: https://trade-nexus-api.metabilityllc1.workers.dev
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        status:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error details
      required:
        - error
...
    PaginationParams:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
          description: Page number
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
          description: Number of items per page
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Maximum number of items to return
        offset:
          type: integer
          minimum: 0
          description: Number of items to skip
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        trade_date:
          type: string
          format: date
        contract_code:
          type: string
        quantity:
          type: integer
          description: Transaction quantity (positive for buy, negative for sell)
        price:
          type: number
          format: float
        trader_id:
          type: integer
        company_id:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
      required:
        - id
        - trade_date
        - contract_code
        - quantity
...
security:
  - BearerAuth: []
paths:
...
  /api/transactions:
    get:
      tags:
        - Position Limits
      summary: List transactions
      description: Retrieve all transactions with optional filtering
      parameters:
        - name: contract_code
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  meta:
                    $ref: '#/components/schemas/PaginationParams'
    post:
      tags:
        - Position Limits
      summary: Create transaction
      description: Create a new trading transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                trade_date:
                  type: string
                  format: date
                contract_code:
                  type: string
                quantity:
                  type: integer
                price:
                  type: number
                  format: float
              required:
                - trade_date
                - contract_code
                - quantity
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid transaction data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Alerts Endpoints
...